# frontend-app/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/dev/create-db-and-tables":{"post":{"tags":["Development"],"summary":"Create Db And Tables","operationId":"create_db_and_tables_api_dev_create_db_and_tables_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/workers/":{"get":{"tags":["Workers"],"summary":"Get Workers","operationId":"getWorkers","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WorkerResponse"},"type":"array","title":"Response Getworkers"}}}}}},"post":{"tags":["Workers"],"summary":"Create Worker","operationId":"createWorker","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkerCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkerResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WorkerCreate":{"properties":{"name":{"type":"string","title":"Name"},"parental_surname":{"type":"string","title":"Parental Surname"},"maternal_surname":{"type":"string","title":"Maternal Surname"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["name","parental_surname","maternal_surname"],"title":"WorkerCreate"},"WorkerResponse":{"properties":{"name":{"type":"string","title":"Name"},"parental_surname":{"type":"string","title":"Parental Surname"},"maternal_surname":{"type":"string","title":"Maternal Surname"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","parental_surname","maternal_surname","id"],"title":"WorkerResponse"}}}}