# frontend-app/src/api/fastAPI.schemas.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
export interface HTTPValidationError {
  detail?: ValidationError[];
}

export type ValidationErrorLocItem = string | number;

export interface ValidationError {
  loc: ValidationErrorLocItem[];
  msg: string;
  type: string;
}

export interface WorkerCreate {
  name: string;
  parental_surname: string;
  maternal_surname: string;
  created_at?: string;
  updated_at?: string;
}

export interface WorkerResponse {
  name: string;
  parental_surname: string;
  maternal_surname: string;
  created_at?: string;
  updated_at?: string;
  id: number;
}